CL-Sort Sorting Package: * Update module_sort.py * Update source_set_sort.py - update docs, update imports * Move changelist_sort/string_operations.py to changelist_sort/sorting/string_operations.py
